+++ src/components/Header.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 import styles from "../styles/Header.module.css"
 export default function Header() {
   return (
